+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/ImpTest.hs
@@ -36,6 +36,7 @@ module Test.Cardano.Ledger.Conway.ImpTest (
   mkConstitutionProposal,
   mkProposal,
   mkProposalWithRewardAccount,
+  mkTreasuryWithdrawalsGovAction,
   submitTreasuryWithdrawals,
   submitVote,
   submitVote_,
@@ -83,7 +84,6 @@ module Test.Cardano.Ledger.Conway.ImpTest (
   getProposalsForest,
   logProposalsForest,
   logProposalsForestDiff,
-  constitutionShouldBe,
   getCCExpiry,
   ccShouldBeExpired,
   ccShouldNotBeExpired,
@@ -92,6 +92,8 @@ module Test.Cardano.Ledger.Conway.ImpTest (
   getLastEnactedCommittee,
   getLastEnactedConstitution,
   submitParameterChange,
+  mkMinFeeUpdateGovAction,
+  mkParameterChangeGovAction,
   mkUpdateCommitteeProposal,
   submitUpdateCommittee,
   expectCommitteeMemberPresence,
@@ -187,7 +189,7 @@ import Cardano.Ledger.Credential (Credential (..), StakeReference (..))
 import Cardano.Ledger.Crypto (Crypto (..))
 import Cardano.Ledger.DRep
 import Cardano.Ledger.Keys (KeyHash, KeyRole (..))
-import Cardano.Ledger.Plutus.Language (Language (..), SLanguage (..))
+import Cardano.Ledger.Plutus.Language (Language (..), SLanguage (..), hashPlutusScript)
 import Cardano.Ledger.PoolParams (ppRewardAccount)
 import qualified Cardano.Ledger.Shelley.HardForks as HardForks (bootstrapPhase)
 import Cardano.Ledger.Shelley.LedgerState (
@@ -233,7 +235,6 @@ import Data.Maybe (fromJust, fromMaybe, isJust)
 import Data.Sequence.Strict (StrictSeq (..))
 import qualified Data.Sequence.Strict as SSeq
 import qualified Data.Set as Set
-import qualified Data.Text as T
 import Data.Tree
 import qualified GHC.Exts as GHC (fromList)
 import Lens.Micro
@@ -245,6 +246,7 @@ import Test.Cardano.Ledger.Core.KeyPair (KeyPair (..), mkCred)
 import Test.Cardano.Ledger.Core.Rational (IsRatio (..))
 import Test.Cardano.Ledger.Imp.Common
 import Test.Cardano.Ledger.Plutus (testingCostModel)
+import Test.Cardano.Ledger.Plutus.Guardrail (guardrailScript)
 
 -- | Modify the PParams in the current state with the given function
 conwayModifyPParams ::
@@ -283,6 +285,7 @@ instance
           { anchorUrl = errorFail $ textToUrl 128 "https://cardano-constitution.crypto"
           , anchorDataHash = hashAnchorData (AnchorData "Cardano Constitution Content")
           }
+      guardrailScriptHash = hashPlutusScript guardrailScript
     pure
       ConwayGenesis
         { cgUpgradePParams =
@@ -290,22 +293,22 @@ instance
               { ucppPoolVotingThresholds =
                   PoolVotingThresholds
                     { pvtMotionNoConfidence = 51 %! 100
-                    , pvtCommitteeNormal = 51 %! 100
-                    , pvtCommitteeNoConfidence = 51 %! 100
+                    , pvtCommitteeNormal = 65 %! 100
+                    , pvtCommitteeNoConfidence = 65 %! 100
                     , pvtHardForkInitiation = 51 %! 100
                     , pvtPPSecurityGroup = 51 %! 100
                     }
               , ucppDRepVotingThresholds =
                   DRepVotingThresholds
                     { dvtMotionNoConfidence = 51 %! 100
-                    , dvtCommitteeNormal = 51 %! 100
-                    , dvtCommitteeNoConfidence = 51 %! 100
-                    , dvtUpdateToConstitution = 51 %! 100
+                    , dvtCommitteeNormal = 65 %! 100
+                    , dvtCommitteeNoConfidence = 65 %! 100
+                    , dvtUpdateToConstitution = 65 %! 100
                     , dvtHardForkInitiation = 51 %! 100
                     , dvtPPNetworkGroup = 51 %! 100
                     , dvtPPEconomicGroup = 51 %! 100
                     , dvtPPTechnicalGroup = 51 %! 100
-                    , dvtPPGovGroup = 51 %! 100
+                    , dvtPPGovGroup = 75 %! 100
                     , dvtTreasuryWithdrawal = 51 %! 100
                     }
               , ucppCommitteeMinSize = 1
@@ -318,7 +321,7 @@ instance
               , -- TODO: Replace with correct cost model.
                 ucppPlutusV3CostModel = testingCostModel PlutusV3
               }
-        , cgConstitution = Constitution constitutionAnchor SNothing
+        , cgConstitution = Constitution constitutionAnchor (SJust guardrailScriptHash)
         , cgCommittee = committee
         , cgDelegs = mempty
         , cgInitialDReps = mempty
@@ -851,6 +854,13 @@ submitGovActions gas = do
   let txId = txIdTx tx
   pure $ NE.zipWith (\idx _ -> GovActionId txId (GovActionIx idx)) (0 NE.:| [1 ..]) gas
 
+mkTreasuryWithdrawalsGovAction ::
+  ConwayEraGov era =>
+  [(RewardAccount (EraCrypto era), Coin)] ->
+  ImpTestM era (GovAction era)
+mkTreasuryWithdrawalsGovAction wdrls =
+  TreasuryWithdrawals (Map.fromList wdrls) <$> getGovPolicy
+
 submitTreasuryWithdrawals ::
   ( ShelleyEraImp era
   , ConwayEraTxBody era
@@ -858,9 +868,8 @@ submitTreasuryWithdrawals ::
   ) =>
   [(RewardAccount (EraCrypto era), Coin)] ->
   ImpTestM era (GovActionId (EraCrypto era))
-submitTreasuryWithdrawals wdrls = do
-  policy <- getGovPolicy
-  submitGovAction $ TreasuryWithdrawals (Map.fromList wdrls) policy
+submitTreasuryWithdrawals wdrls =
+  mkTreasuryWithdrawalsGovAction wdrls >>= submitGovAction
 
 enactTreasuryWithdrawals ::
   ConwayEraImp era =>
@@ -880,9 +889,23 @@ submitParameterChange ::
   StrictMaybe (GovActionId (EraCrypto era)) ->
   PParamsUpdate era ->
   ImpTestM era (GovActionId (EraCrypto era))
-submitParameterChange parent ppu = do
-  policy <- getGovPolicy
-  submitGovAction $ ParameterChange (GovPurposeId <$> parent) ppu policy
+submitParameterChange parent ppu =
+  mkParameterChangeGovAction parent ppu >>= submitGovAction
+
+mkParameterChangeGovAction ::
+  ConwayEraImp era =>
+  StrictMaybe (GovActionId (EraCrypto era)) ->
+  PParamsUpdate era ->
+  ImpTestM era (GovAction era)
+mkParameterChangeGovAction parent ppu =
+  ParameterChange (GovPurposeId <$> parent) ppu <$> getGovPolicy
+
+mkMinFeeUpdateGovAction ::
+  ConwayEraImp era =>
+  StrictMaybe (GovActionId (EraCrypto era)) -> ImpTestM era (GovAction era)
+mkMinFeeUpdateGovAction p = do
+  minFeeValue <- uniformRM (30, 1000)
+  mkParameterChangeGovAction p (def & ppuMinFeeAL .~ SJust (Coin minFeeValue))
 
 getGovPolicy :: ConwayEraGov era => ImpTestM era (StrictMaybe (ScriptHash (EraCrypto era)))
 getGovPolicy =
@@ -1491,14 +1514,6 @@ enactConstitution prevGovId constitution dRep committeeMembers = impAnn "Enactin
   enactedConstitution <- getsNES $ newEpochStateGovStateL . constitutionGovStateL
   enactedConstitution `shouldBe` constitution
   pure govId
-
--- | Asserts that the URL of the current constitution is equal to the given
--- string
-constitutionShouldBe :: (HasCallStack, ConwayEraGov era) => String -> ImpTestM era ()
-constitutionShouldBe cUrl = do
-  Constitution {constitutionAnchor = Anchor {anchorUrl}} <-
-    getsNES $ newEpochStateGovStateL . constitutionGovStateL
-  anchorUrl `shouldBe` errorFail (textToUrl 128 $ T.pack cUrl)
 
 expectNumDormantEpochs :: HasCallStack => EpochNo -> ImpTestM era ()
 expectNumDormantEpochs expected = do